Estimate states in a scalar stochastic differential equation based on discretization to a HMM

HMMfilterSDE computes the state estimation for a discretely observed stochastic differential equation

HMMfilterSDE(
  u,
  D,
  xi,
  bc,
  x0dist,
  tvec,
  yvec,
  lfun,
  do.smooth = FALSE,
  do.Viterbi = FALSE,
  pfun = NULL
)

Arguments

u

function mapping state (numeric scalar) to advective term (numeric scalar)

D

function mapping state (numeric scalar) to diffusivity (numeric scalar)

xi

The numerical grid. Numeric vector of increasing values, giving cell boundaries

bc

String indicating boundary conditions. See details.

x0dist

Prior distribution of the initial state, specified as a c.d.f. or as a vector of same length as xi, start

tvec

Vector of (increasing) time points where observations are taken

yvec

Vector of observations at each time point

lfun

Likelihood function so that lfun(x,y) gives the likelihood of y given x

do.smooth

Do want smoothing, or only predictive filtering / estimation?

do.Viterbi

Do we want the most probable state sequence, found with the Viterbi algorithm?

pfun

C.d.f. of observations given states, i.e. pfun(x,y) gives P(Y<=y | X = x). If supplied, pseudo-prediction residuals will be computed

Value

A list containing: phi A tabulation of the predicitive probability densities psi A tabulation of the estimated probability densities pi (If do.smooth==TRUE) A tabulation of the smoothed probability densities c A vector containing the normalization constants at each time step

Author

Uffe Høgsbro Thygesen